Fixing a -Wsign-compare warning; NFC.
I think it might make sense to make COFF::MaxNumberOfSections16 be a uint32_t, however, that may have wider-reaching implications in other projects, which is why I did not change that declaration.

llvm-svn: 220384

+++ llvm/tools/yaml2obj/yaml2coff.cpp
@@ -37,7 +37,8 @@ struct COFFParser {
   }
 
   bool useBigObj() const {
-    return Obj.Sections.size() > COFF::MaxNumberOfSections16;
+    return static_cast<int32_t>(Obj.Sections.size()) >
+           COFF::MaxNumberOfSections16;
   }
 
   unsigned getHeaderSize() const {
